requirements.txt
feedparser
requests
openai>=1.0.0
python-dotenv
